Got text decorations somewhat working. Split up files
