+++ Source/SVGImage/SVG/SVGRender.cs
@@ -65,6 +65,15 @@ namespace SVGImage.SVG
             return this.CreateDrawing(this.SVG);
         }
 
+        internal void LoadDrawingWithoutRender(string filename)
+        {
+            this.SVG = new SVG(filename, ExternalFileLoader);
+        }
+        internal void LoadDrawingWithoutRender(Stream stream)
+        {
+            this.SVG = new SVG(stream, ExternalFileLoader);
+        }
+
         public DrawingGroup LoadXmlDrawing(string fileXml)
         {
             this.SVG = new SVG(this.ExternalFileLoader);
@@ -428,16 +437,15 @@ namespace SVGImage.SVG
                     AddDrawingToGroup(grp, shape, i);
                     continue;
                 }
-                if (shape is Text textShape)
+                if (shape is TextShape textShape)
                 {
-                    TextRender2 textRender2 = new TextRender2();
+                    TextRender textRender2 = new TextRender();
                     GeometryGroup gp = textRender2.BuildTextGeometry(textShape);
                     if (gp != null)
                     {
                         foreach (Geometry gm in gp.Children)
                         {
-                            TextSpan tspan = TextRender2.GetElement(gm);
-                            if (tspan != null)
+                            if (TextRenderBase.GetElement(gm) is TextShapeBase tspan)
                             {
                                 var di = this.NewDrawingItem(tspan, gm);
                                 AddDrawingToGroup(grp, shape, di);
