+++ Source/SVGImage/SVG/Shapes/Group.cs
@@ -89,7 +89,7 @@ namespace SVGImage.SVG.Shapes
                 retVal = new AnimateTransform(svg, childnode, parent);
             else if (nodeName == SVGTags.sText)
                 //retVal = new TextShape(svg, childnode, parent);
-                retVal = new Text(svg, childnode, parent);
+                retVal = new TextShape(svg, childnode, parent);
             else if (nodeName == SVGTags.sLinearGradient)
             {
                 svg.PaintServers.Create(svg, childnode);
