+++ Source/SVGImage/SVG/Shapes/ITextChild.cs
@@ -1,0 +1,11 @@
+﻿namespace SVGImage.SVG.Shapes
+{
+    public interface ITextChild : ITextNode
+    {
+        Shape Parent { get; set; }
+    }
+
+
+
+
+}
