+++ Source/SVGImage/SVG/Shapes/ITextNode.cs
@@ -1,0 +1,15 @@
+﻿namespace SVGImage.SVG.Shapes
+{
+    public interface ITextNode
+    {
+        CharacterLayout GetFirstCharacter();
+        CharacterLayout GetLastCharacter();
+        string GetText();
+        int GetLength();
+        CharacterLayout[] GetCharacters();
+    }
+
+
+
+
+}
