+++ Source/SVGImage/SVG/Shapes/LengthOrientation.cs
@@ -1,0 +1,13 @@
+﻿namespace SVGImage.SVG.Shapes
+{
+    public enum LengthOrientation
+    {
+        None,
+        Horizontal,
+        Vertical,
+    }
+
+
+
+
+}
