+++ Source/SVGImage/SVG/Shapes/PolygonShape.cs
@@ -8,15 +8,9 @@ namespace SVGImage.SVG.Shapes
 
     public sealed class PolygonShape : Shape
     {
-        private static Fill DefaultFill = null;
-
         public PolygonShape(SVG svg, XmlNode node)
             : base(svg, node)
         {
-            if (DefaultFill == null)
-            {
-                DefaultFill = Fill.CreateDefault(svg, "black");
-            }
 
             string points = XmlUtil.AttrValue(node, SVGTags.sPoints, string.Empty);
             var split = new StringSplitter(points);
@@ -30,14 +24,9 @@ namespace SVGImage.SVG.Shapes
 
         public Point[] Points { get; private set; }
 
-        public override Fill Fill
+        protected override Fill DefaultFill()
         {
-            get
-            {
-                Fill f = base.Fill;
-                if (f == null) f = DefaultFill;
-                return f;
-            }
+            return Fill.CreateDefault(Svg, "black");
         }
     }
 }
