+++ Source/SVGImage/SVG/Shapes/TextShape.cs
@@ -1,0 +1,17 @@
+﻿using System.Xml;
+
+namespace SVGImage.SVG.Shapes
+{
+    public class TextShape : TextShapeBase
+    {
+        public TextShape(SVG svg, XmlNode node, Shape parent) : base(svg, node, parent)
+        {
+            
+        }
+
+    }
+
+
+
+
+}
