+++ Source/SVGImage/SVG/Shapes/TextSpan.cs
@@ -1,0 +1,17 @@
+﻿using System.Xml;
+
+namespace SVGImage.SVG.Shapes
+{
+    public class TextSpan : TextShapeBase, ITextChild
+    {
+
+        public TextSpan(SVG svg, XmlNode node, Shape parent) : base(svg, node, parent)
+        {
+        }
+
+    }
+
+
+
+
+}
